back/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn==0.24.0

# Data validation and serialization
pydantic==2.7.1
python-dateutil==2.9.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# HTTP and API clients
requests==2.32.3
httpx==0.27.0
aiohttp==3.12.15

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-mock==3.14.0

# Development tools
black==24.4.2
ruff==0.4.5
mypy==1.10.0
pylint==3.2.2

# Utilities
python-dotenv==1.0.1
pytz==2024.1

# Observability and monitoring
prometheus-client==0.19.0

# Job scheduling
apscheduler==3.10.4

# Authentication and security
pyjwt[crypto]==2.8.0
argon2-cffi==23.1.0
python-multipart==0.0.9